Move events to cog
